refactor(W2.1): split mcp/handlers/instance.rs by concept; end the size-driven era
Pure structural extraction, zero API change, zero behavior change. instance.rs
(748/750 cap) split by concept into sibling modules; instance.rs becomes a
17-line re-export harness so every `instance::handle_*` / `instance::resolve_team_layout`
caller (dispatch adapters, mod.rs, tests) is unchanged.

- instance_queries.rs (75 LOC) — read-only: list / describe.
- instance_state/ (directory module = ONE concept) — lifecycle:
    mod.rs (449)       create / delete / start / replace / restart
                       + restart_spawn_params + resolve_team_layout + tests
    spawn.rs (291)     was instance_spawn.rs  (git rename)
    lifecycle.rs (438) was instance_lifecycle.rs (git rename)
    lifecycle_tests.rs was instance_lifecycle/tests.rs (git rename)
  spawn + lifecycle were size-driven top-level files; they are now legitimate
  submodules of the state concept, ending the "extracted for file_size_invariant"
  apologetic module-doc comments (#964 / Sprint 54 P1-B).
- instance_metadata.rs (231 LOC) — per-instance attributes & control:
  display-name / description / waiting-on / interrupt / move_pane / pane_snapshot
  / report_health / clear_blocked_reason.

Function bodies moved byte-for-byte (git rename for the two whole files; awk
range-extraction for the in-file splits — normalized-diff clean). Only deltas
beyond the moves: trimmed per-module import headers; internal refs follow the
move (spawn::/lifecycle::/super::resolve_team_layout); spawn_single_instance_impl
visibility pub(super)→pub(in crate::mcp::handlers) to preserve the instance_964
test's exact reach; external callers' paths updated (teams.rs, app/overlay.rs →
instance_state::lifecycle::full_delete_instance; instance_964_tests →
instance_state::spawn); stale filename refs swept (tools.rs provenance docs,
comms_p0c_tests doc).

Three concepts share no state (per survey). Verified: clippy --all-targets
--features tray -D warnings clean; file_size_invariant green (all split files
< 750); 437 mcp::handlers + 159 instance + 42 instance_964/lifecycle + 45
mcp::tools tests pass, no assertion changes.


Agend-Agent: fixup-dev-2
Agend-Task: t-20260612105918317475-27
Agend-Branch: refactor/w2-1-instance-split
Agend-Issued-At: 2026-06-12T10:59:35.483296+00:00
