+++ src/mcp/handlers/instance_state/lifecycle_tests.rs
