+++ src/mcp/handlers/mod.rs
@@ -10,8 +10,9 @@ pub(crate) mod dispatch_hook;
 pub(crate) mod evidence_gate;
 mod force_release;
 pub(crate) mod instance;
-pub(crate) mod instance_lifecycle;
-mod instance_spawn;
+mod instance_metadata;
+mod instance_queries;
+pub(crate) mod instance_state;
 mod restart;
 mod schedule;
 pub(crate) mod sha_gate;
